+++ packages/react/src/components/Typography/Body/index.ts
@@ -1,1 +1,1 @@
-export { default as BodyLong } from './BodyLong';
+export { default as BodyLong } from './Body';
